atualização de branch

+++ src/main/java/br/com/cefet/controller/ManutencaoController.java
@@ -20,6 +20,8 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 import org.springframework.web.servlet.ModelAndView;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
@@ -28,14 +30,23 @@ import br.com.cefet.dto.requisicaoManutencao;
 import br.com.cefet.dto.requisicaoReserva;
 import br.com.cefet.dto.requisicaoVeiculo;
 import br.com.cefet.model.Categoria;
+import br.com.cefet.model.Cliente;
+import br.com.cefet.model.Contrato;
 import br.com.cefet.model.Estacao;
+import br.com.cefet.model.Funcionario;
 import br.com.cefet.model.Marca;
+import br.com.cefet.model.Motorista;
 import br.com.cefet.model.Paletas;
+import br.com.cefet.model.Status;
+import br.com.cefet.model.Usuario;
 import br.com.cefet.model.Manutencao;
 import br.com.cefet.model.Veiculo;
 import br.com.cefet.repository.EstacaoRepository;
 import br.com.cefet.repository.ManutencaoRepository;
+import br.com.cefet.repository.UsuarioRepository;
 import br.com.cefet.repository.VeiculoRepository;
+import br.com.cefet.service.SessaoService;
+import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 
 @Controller
@@ -47,20 +58,47 @@ public class ManutencaoController {
 	@Autowired
 	private VeiculoRepository veiculoRepository;
 
+	@Autowired
+	private UsuarioRepository usuarioRepository;
+
+	@Autowired
+	private SessaoService sessaoService;
+
 	@GetMapping("/manutencoes")
 	public ModelAndView index() {
 		ModelAndView mv = new ModelAndView("manutencoes/index");
-
-		List<Manutencao> manutencoes = this.manutencaoRepository.findAll();
-		mv.addObject("manutencoes", manutencoes);
-
-		return mv;
+		HttpSession session = ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest()
+				.getSession();
+		Cliente cliente = sessaoService.obterClienteDaSessao(session);
+		if (cliente != null) {
+			List<Manutencao> manutencoes = this.manutencaoRepository.findByStatusAndUsuario("Corrente", cliente);
+			mv.addObject("manutencoes", manutencoes);
+			mv.addObject("usuario", cliente);
+			return mv;
+		} else {
+			Funcionario funcionario = sessaoService.obterFuncionarioDaSessao(session);
+			if (funcionario != null) {
+				List<Manutencao> manutencoes = this.manutencaoRepository.findByStatus("Corrente");
+				mv.addObject("manutencoes", manutencoes);
+				mv.addObject("usuario", funcionario);
+				return mv;
+			}
+		}
+		System.out.println("Não há contratos acossiados!");
+		return new ModelAndView("redirect:/sessoes");
 	}
 
 	@GetMapping("/manutencoes/new")
 	public ModelAndView novo(@RequestParam(name = "veiculoId") int veiculoId) {
 		Optional<Veiculo> optionalVeiculo = veiculoRepository.findById(veiculoId);
-
+		HttpSession session = ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest()
+				.getSession();
+		Cliente cliente = sessaoService.obterClienteDaSessao(session);
+		Funcionario funcionario = sessaoService.obterFuncionarioDaSessao(session);
+		if (cliente == null && funcionario == null) {
+			System.out.println("Não há sessão associada!");
+			return new ModelAndView("redirect:/sessoes");
+		}
 		if (optionalVeiculo.isPresent()) {
 			Veiculo veiculo = optionalVeiculo.get();
 
@@ -76,6 +114,33 @@ public class ManutencaoController {
 			return new ModelAndView("redirect:/veiculos");
 		}
 	}
+	
+	
+	@GetMapping("/manutencoes/archive")
+	public ModelAndView historico() {
+		ModelAndView mv = new ModelAndView("manutencoes/archive");
+		HttpSession session = ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest().getSession();
+		Cliente cliente = sessaoService.obterClienteDaSessao(session);
+		if (cliente != null) {
+			List<Manutencao> manutencoes = this.manutencaoRepository.findByStatusAndUsuario("Arquivado", cliente);
+			
+			mv.addObject("cliente", cliente);
+			mv.addObject("manutencoes", manutencoes);
+			return mv;
+			} 
+			else {
+				Funcionario funcionario = sessaoService.obterFuncionarioDaSessao(session);
+				if (funcionario != null) {
+					List<Manutencao> manutencoes = this.manutencaoRepository.findByStatus("Arquivado");
+					
+					mv.addObject("funcionario", funcionario);
+					mv.addObject("manutencoes", manutencoes);
+					return mv;
+			}
+			System.out.println("Não há contratos acossiados!");
+			return new ModelAndView("redirect:/sessoes");
+		}
+	}
 
 //	 O bloco abaixo cria um objeto requisicaoVeiculo para tratar erro de validação
 //	 de dados thymeleaf
@@ -85,9 +150,19 @@ public class ManutencaoController {
 	}
 
 	@PostMapping("/manutencoes")
-	public ModelAndView create(@Valid requisicaoManutencao requisicao, BindingResult result, RedirectAttributes redirectAttributes) {
+	public ModelAndView create(@Valid requisicaoManutencao requisicao, BindingResult result,
+			RedirectAttributes redirectAttributes) {
 		Veiculo veiculo = veiculoRepository.findById(requisicao.getVeiculoId()).orElse(null);
 		System.out.printf("ID VEICULO - %d%n", requisicao.getVeiculoId());
+		HttpSession session = ((ServletRequestAttributes) RequestContextHolder.currentRequestAttributes()).getRequest()
+				.getSession();
+		Cliente cliente = sessaoService.obterClienteDaSessao(session);
+		Funcionario funcionario = sessaoService.obterFuncionarioDaSessao(session);
+		Usuario usuario = cliente;
+		if (cliente == null) {
+			usuario = funcionario;
+		}
+
 		int estacaoId = 0;
 		try {
 			estacaoId = Integer.parseInt(requisicao.getEstacaoId());
@@ -99,9 +174,8 @@ public class ManutencaoController {
 		System.out.printf("ID ESTACAO - %d%n", estacaoId);
 		ModelAndView mv = new ModelAndView("manutencoes/new");
 
-		
-		 List<Estacao> estacoes = estacaoRepository.findAll();
-		 mv.addObject("estacoes", estacoes);
+		List<Estacao> estacoes = estacaoRepository.findAll();
+		mv.addObject("estacoes", estacoes);
 		if (result.hasErrors()) {
 			System.out.println("\n**********************Invalid Input Found**************************\n");
 
@@ -124,17 +198,21 @@ public class ManutencaoController {
 			}
 			mv.addObject("veiculo", veiculo);
 			mv.addObject("estacao", estacao);
+			mv.addObject("usuario", usuario);
 			mv.addObject("dataEntrada", requisicao.getDataEntrada());
 			mv.addObject("dataSaida", requisicao.getDataSaida());
 			return mv;
 		} else {
-		if (veiculo != null && estacao != null && !veiculoRepository.isVeiculoEmManutencao(veiculo.getId())) {
+			if (veiculo != null && estacao != null && !veiculoRepository.isVeiculoEmManutencao(veiculo.getId())) {
 				Manutencao manutencao = new Manutencao();
 				mv = new ModelAndView("redirect:/manutencoes/" + manutencao.getIdManutencao());
 
 				manutencao.setEstacao(estacao);
 				manutencao.setVeiculo(veiculo);
+				manutencao.setUsuario(usuario);
+				manutencao.setStatus("Corrente");
 				veiculo.setEmManutencao(true);
+				estacao.setStatus(Status.Reservado);
 				manutencao = requisicao.toManutencao(manutencao);
 
 				System.out.println("Data de entrada recebida: " + manutencao.getDataEntrada());
@@ -178,14 +256,60 @@ public class ManutencaoController {
 		}
 	}
 
-	@GetMapping("/manutencoes/{idManutencao}/delete")
-	public String delete(@PathVariable Integer idManutencao) {
+	@GetMapping("/manutencoes/{idManutencao}/archive")
+	public String archive(@PathVariable Integer idManutencao) {
 		try {
-			this.manutencaoRepository.deleteById(idManutencao);
-			return "redirect:/manutencoes";
+			Optional<Manutencao> optional = this.manutencaoRepository.findById(idManutencao);
+			if (optional.isPresent()) {
+				Manutencao manutencao = optional.get();
+				Estacao estacao = manutencao.getEstacao();
+				Veiculo veiculo = manutencao.getVeiculo();
+				// Atualiza o status da estação para 'Livre'
+				manutencao.setStatus("Arquivado");
+				this.manutencaoRepository.save(manutencao);
+				estacao.setStatus(Status.Livre);
+				estacaoRepository.save(estacao);
+				veiculo.setEmManutencao(false);
+				veiculoRepository.save(veiculo);
+
+				return "redirect:/manutencoes";
+			} else {
+				System.out
+						.println("Registro não consta no banco ou não foi encontrado, portanto não pode ser deletado.");
+				return "redirect:/manutencoes";
+			}
 		} catch (EmptyResultDataAccessException e) {
 			System.out.println("Registro não consta no banco ou não foi encontrado, portanto não pode ser deletado.");
 			return "redirect:/manutencoes";
 		}
 	}
+
+	@GetMapping("/manutencoes/{idManutencao}/delete")
+	public String delete(@PathVariable Integer idManutencao) {
+		try {
+			Optional<Manutencao> optional = this.manutencaoRepository.findById(idManutencao);
+			if (optional.isPresent()) {
+				Manutencao manutencao = optional.get();
+				Estacao estacao = manutencao.getEstacao();
+				Veiculo veiculo = manutencao.getVeiculo();
+				this.manutencaoRepository.deleteById(idManutencao);
+
+				// Atualiza o status da estação para 'Livre'
+				estacao.setStatus(Status.Livre);
+				estacaoRepository.save(estacao);
+				veiculo.setEmManutencao(false);
+				veiculoRepository.save(veiculo);
+
+				return "redirect:/manutencoes";
+			} else {
+				System.out
+						.println("Registro não consta no banco ou não foi encontrado, portanto não pode ser deletado.");
+				return "redirect:/manutencoes";
+			}
+		} catch (EmptyResultDataAccessException e) {
+			System.out.println("Registro não consta no banco ou não foi encontrado, portanto não pode ser deletado.");
+			return "redirect:/manutencoes";
+		}
+	}
+
 }
